added viewModel func

+++ Calculator/CalculateViewModel.swift
@@ -10,25 +10,25 @@ import Foundation
 class CalculateViewModel {
     
     var result: Int = 0
-    var number1: Double = 0
-    var number2: Double = 0
+    var number1: Int = 0
+    var number2: Int = 0
      
-     func add(_ number1: Double, _ number2: Double)  -> Int{
+     func add(_ number1: Int, _ number2: Int)  -> Int{
         return Int(number1 + number2)
          
      }
     
-    func minus(_ number1: Double, _ number2: Double)  -> Int{
+    func minus(_ number1: Int, _ number2: Int)  -> Int{
        return Int(number1 - number2)
         
     }
     
-    func divide(_ number1: Double, _ number2: Double)  -> Int{
+    func divide(_ number1: Int, _ number2: Int)  -> Int{
        return Int(number1 / number2)
         
     }
     
-    func multiply(_ number1: Double, _ number2: Double)  -> Int{
+    func multiply(_ number1: Int, _ number2: Int)  -> Int{
        return Int(number1 * number2)
         
     }
